feat(DangerState): implement danger state management and update device response

+++ src/main/resources/db/migration/V20260611_1__add_device_resource_usage.sql
@@ -1,3 +1,3 @@
 ALTER TABLE device
-    ADD COLUMN cpu_usage    DECIMAL(5, 2) NULL COMMENT 'CPU 사용률 (%)',
-    ADD COLUMN memory_usage DECIMAL(5, 2) NULL COMMENT '메모리 사용률 (%)';
+    ADD COLUMN cpu_usage    DOUBLE NULL COMMENT 'CPU 사용률 (%)',
+    ADD COLUMN memory_usage DOUBLE NULL COMMENT '메모리 사용률 (%)';
